before use manal optimization

+++ src/csi_sign_language/modules/loss.py
@@ -19,8 +19,8 @@ class VACLoss:
         #[t, n, c] logits
         conv_out, seq_out = F.log_softmax(conv_out, dim=-1), F.log_softmax(seq_out, dim=-1)
 
-        conv_out = conv_out.to(torch.float32)
-        seq_out = seq_out.to(torch.float32)
+        # conv_out = conv_out.to(torch.float32)
+        # seq_out = seq_out.to(torch.float32)
 
         loss = 0
         if self.weights[0] > 0.:
